requirements.txt
# Core data processing
pandas==2.2.3
numpy==2.2.5
pyarrow==19.0.1

# Database
mysql-connector-python==9.3.0

# HTTP requests
requests==2.32.3

# Date handling
python-dateutil==2.9.0.post0
pytz==2025.2

# Environment variables
python-dotenv==1.0.1
